Rename common/scripting to common/lua
