+++ src/server/lua/data_api.cpp
@@ -1,6 +1,6 @@
 #include "server_lua_api.h"
 
-#include <common/scripting/script_engine.h>
+#include <common/lua/script_engine.h>
 #include <common/world/biome.h>
 #include <common/world/voxel_data.h>
