+++ src/server/lua/util_api.cpp
@@ -1,5 +1,5 @@
 #include "server_lua_api.h"
-#include <common/scripting/script_engine.h>
+#include <common/lua/script_engine.h>
 #include <common/util/random_number_generator.h>
 
 namespace {
