+++ src/server/network/server.h
@@ -5,7 +5,7 @@
 #include <array>
 #include <common/network/command_dispatcher.h>
 #include <common/network/net_host.h>
-#include <common/scripting/script_engine.h>
+#include <common/lua/script_engine.h>
 #include <common/world/biome.h>
 #include <common/world/chunk_manager.h>
 #include <common/world/voxel_data.h>
